Added test for Recipe class method and cleaned up duplicate files in project directory
