--- Reciplease/RecipleaseTests/JSONs/RecipeDataEmpty 2.json
@@ -1,7 +1,0 @@
-{
-    "from": 1,
-    "to": 0,
-    "count": 0,
-    "_links": {},
-    "hits": []
-}
